requirements.txt
alabaster==0.7.12
argh==0.26.2
Babel==2.7.0
bumpversion==0.5.3
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.1
Django==2.2.3
djangorestframework==3.10.2
dlib==19.17.99
docutils==0.15.2
face-recognition==1.2.3
face-recognition-models==0.3.0
flake8==2.6.0
Flask==1.1.1
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
mccabe==0.5.3
numpy==1.17.0
opencv-contrib-python==4.1.0.25
pathtools==0.1.2
Pillow==6.1.0
pluggy==0.3.1
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.0.0
pycparser==2.19
pyflakes==1.2.3
Pygments==2.4.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==1.4.8
sqlparse==0.3.0
tox==2.3.1
urllib3==1.25.3
virtualenv==16.7.2
watchdog==0.8.3
Werkzeug==0.15.5
